# webAPI/bin/Debug/net6.0/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "webAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/ToDo": {
      "get": {
        "tags": [
          "ToDo"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/myitem"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/myitem"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/myitem"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ToDo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/myitem"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/myitem"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/myitem"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/ToDo/today-items": {
      "get": {
        "tags": [
          "ToDo"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/myitem"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/myitem"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/myitem"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/ToDo/check/{id}": {
      "get": {
        "tags": [
          "ToDo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/ToDo/{tid}": {
      "get": {
        "tags": [
          "ToDo"
        ],
        "parameters": [
          {
            "name": "tid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/myitem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/myitem"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/myitem"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/ToDo/{itemid}": {
      "delete": {
        "tags": [
          "ToDo"
        ],
        "parameters": [
          {
            "name": "itemid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/ToDo/update": {
      "post": {
        "tags": [
          "ToDo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/myitem"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/myitem"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/myitem"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/ToDo/make-done": {
      "post": {
        "tags": [
          "ToDo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/myitem"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/myitem"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/myitem"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "myitem": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "myitemId": {
            "maxLength": 5,
            "minLength": 1,
            "type": "string",
            "nullable": true
          },
          "dateToDone": {
            "type": "string",
            "format": "date-time"
          },
          "isDone": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      }
    }
  }
}